Handle the case when PWD is not set.

+++ interfaces/daqp-python/setup.py
@@ -7,9 +7,12 @@ import pathlib
 
 
 # Copy C source
-src_path = pathlib.Path(os.environ["PWD"], "../../../daqp")
+try:
+    src_path = pathlib.Path(os.environ["PWD"], "../../../daqp")
+except:
+    src_path = []
 csrc_dir = pathlib.Path('./csrc')
-if os.path.exists(src_path) and not os.path.exists(csrc_dir):
+if src_path and os.path.exists(src_path) and not os.path.exists(csrc_dir):
     os.mkdir(csrc_dir)
     copytree(os.path.join(src_path,'src'),os.path.join(csrc_dir,'src'))
     copytree(os.path.join(src_path,'include'),os.path.join(csrc_dir,'include'))
@@ -87,6 +90,6 @@ setup(name='daqp',
         zip_safe=False)
 
 # Cleanup C-source
-if os.path.exists(src_path):
+if src_path and os.path.exists(src_path):
     rmtree(csrc_dir)
     os.remove(pathlib.Path('./LICENSE'))
